requirements.txt
chromadb
fastapi
uvicorn
pandas
rank-bm25
openai
python-dotenv
tqdm
streamlit
rouge-score
nltk

